Merge origin

+++ application/GW2 Addon Manager/Backend/Addons/ApprovedList.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.ObjectModel;
+﻿using System;
+using System.Collections.ObjectModel;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Net;
 using GW2_Addon_Manager.App.Configuration;
+using GW2_Addon_Manager.Backend;
 using Newtonsoft.Json;
 
 namespace GW2_Addon_Manager
@@ -11,6 +13,7 @@ namespace GW2_Addon_Manager
     class ApprovedList
     {
         private const string AddonFolder = "resources\\addons";
+
         //Approved-addons repository
         private const string RepoUrl = "https://api.github.com/repositories/206052865";
 
@@ -24,19 +27,18 @@ namespace GW2_Addon_Manager
         /// <summary>
         /// Check current version of addon list against remote repo for changes and fetch them
         /// </summary>
-        public void FetchListFromRepo()
+        private void FetchListFromRepo()
         {
             const string tempFileName = "addonlist";
-            var client = new WebClient();
-            client.Headers.Add("User-Agent", "Gw2 Addon Manager");
-
-            try
+            using (var client = WebClientFactory.Create())
             {
-                var raw = client.DownloadString(RepoUrl + "/branches");
+                var updateHelper = new UpdateHelper(client);
+                var raw = updateHelper.DownloadStringFromGithubAPI(RepoUrl + "/branches");
+
                 var result = JsonConvert.DeserializeObject<BranchInfo[]>(raw);
                 var master = result.Single(r => r.Name == "master").Commit.Sha;
 
-                if (master == _configManager.UserConfig.AddonsList.Hash) return;
+                if (master == _configManager.UserConfig.AddonsList.Hash || master is null) return;
 
                 if (Directory.Exists(AddonFolder))
                     Directory.Delete(AddonFolder, true);
@@ -44,9 +46,8 @@ namespace GW2_Addon_Manager
                     File.Delete(tempFileName);
 
                 //fetching new version
-                client = new WebClient();
-                client.Headers.Add("User-Agent", "Gw2 Addon Manager");
-                client.DownloadFile(RepoUrl + "/zipball", tempFileName);
+                updateHelper.DownloadFileFromGithubAPI(RepoUrl + "/zipball", tempFileName);
+
                 ZipFile.ExtractToDirectory(tempFileName, AddonFolder);
                 var downloaded = Directory.EnumerateDirectories(AddonFolder).First();
                 foreach (var entry in Directory.EnumerateFileSystemEntries(downloaded))
@@ -62,13 +63,6 @@ namespace GW2_Addon_Manager
                 Directory.Delete(downloaded, true);
                 File.Delete(tempFileName);
             }
-            catch (WebException ex)
-            {
-                //TODO Log error
-                if(((HttpWebResponse) ex.Response).StatusCode == HttpStatusCode.Forbidden)
-                    return;
-                throw;
-            }
         }
 
         /// <summary>
@@ -80,7 +74,8 @@ namespace GW2_Addon_Manager
             FetchListFromRepo();
 
             var addons = new ObservableCollection<AddonInfoFromYaml>(); //List of AddonInfo objects
-            var addonDirectories = Directory.GetDirectories(AddonFolder);  //Names of addon subdirectories in /resources/addons
+            var addonDirectories =
+                Directory.GetDirectories(AddonFolder); //Names of addon subdirectories in /resources/addons
             foreach (var addonFolderName in addonDirectories)
             {
                 if (addonFolderName == "resources\\addons\\d3d9_wrapper") continue;
@@ -88,7 +83,7 @@ namespace GW2_Addon_Manager
                 var addonInfo = AddonYamlReader.getAddonInInfo(addonFolderName.Replace(AddonFolder + "\\", ""));
                 addonInfo.folder_name = addonFolderName.Replace(AddonFolder + "\\", "");
                 addonInfo.IsSelected = _configManager.UserConfig.AddonsList[addonInfo.folder_name]?.Installed ?? false;
-                addons.Add(addonInfo);       //retrieving info from each addon subdirectory's update.yaml file and adding it to the list
+                addons.Add(addonInfo); //retrieving info from each addon subdirectory's update.yaml file and adding it to the list
             }
 
             return addons;
